+++ Sources/HIGThemesContract/HIGTheme.swift
@@ -47,6 +47,7 @@ public protocol HIGTheme: Sendable {
     var treeView: any HIGTreeViewTokens { get }
     var reorderableList: any HIGReorderableListTokens { get }
     var calendar: any HIGCalendarTokens { get }
+    var dashboardGrid: any HIGDashboardGridTokens { get }
     var dataTable: any HIGDataTableTokens { get }
     var breadcrumb: any HIGBreadcrumbTokens { get }
     var pageHeader: any HIGPageHeaderTokens { get }